requirements.txt
# 这里装需要的外部库，python源码库不需要
pandas
numpy
scikit-learn==0.24.2
joblib
sklearn
pandas_profiling
streamlit_pandas_profiling